Moved invite accepts to group tab and created partials for invite_count and retro_active.
