+++ app/hindsite/group/group.py
@@ -3,16 +3,18 @@ Template route testing for development
 """
 import os
 from flask import Blueprint, flash, redirect, render_template, request, session, url_for
-from flask_login import login_required
-from app.hindsite.group.group_model import send_invitation
+from flask_login import login_required, current_user
+from app.hindsite.group.group_model import send_invitation, get_invitations, get_invitation, \
+    accept_invitation
 
 from app.hindsite.group.group_model import UserSearchError, get_uninvited_users, get_invited_users
 
 static_dir = os.path.abspath('static')
 group = Blueprint('group',
-                   __name__,
-                   template_folder='templates',    # relative route to templates dir
-                   static_folder=static_dir)
+                  __name__,
+                  template_folder='templates',  # relative route to templates dir
+                  static_folder=static_dir)
+
 
 @group.route('/group', methods=['GET', 'POST'])
 @login_required
@@ -25,6 +27,7 @@ def group_page():
     if group is not None:
         users = get_invited_users(group)
     return render_template('group.html', title='Group', users=users)
+
 
 @group.route('/search-users', methods=['GET', 'POST'])
 @login_required
@@ -42,7 +45,7 @@ def search_users():
                 return render_template('partials/no-group.html')
             # Load in the uninvited users
             search = request.form['search']
-            users = get_uninvited_users(session['groupid'],search)
+            users = get_uninvited_users(session['groupid'], search)
         except UserSearchError as e:
             error = e.message
         if error is not None:
@@ -51,6 +54,7 @@ def search_users():
     if request.form['search'] == "":
         users = ""
     return render_template('partials/search-results.html', users=users, term=search)
+
 
 @group.route('/send-invite', methods=['GET', 'POST'])
 @login_required
@@ -76,3 +80,22 @@ def send_invite():
         except UserSearchError as ex:
             flash(ex.message)
     return render_template('partials/search-results.html')
+
+
+@group.route('/invites', methods=['POST', 'GET'])
+@login_required
+def invites():
+    """
+        Loads all the invite codes to be accepted or rejected
+    TODO: Put the invite codes into a modal, add decline button, create
+    notification bell to open the modal.
+    """
+    error = None
+    if request.method == 'GET':
+        invitations = get_invitations(current_user.id)
+        return render_template('partials/invites.html', invitations=invitations)
+    if request.method == 'POST':
+        group = request.args['group']
+        membership = get_invitation(group, current_user.id)
+        accept_invitation(membership)
+    return render_template('partials/accepted.html')
