+++ app/hindsite/group/group_model.py
@@ -51,6 +51,7 @@ def send_invitation(group_id: int, email: str):
     db.session.commit()
     return membership
 
+
 def get_uninvited_users(group_id: int, term: str):
     """
     Gets all user records that haven't been invited to a group.
@@ -73,6 +74,7 @@ def get_uninvited_users(group_id: int, term: str):
             uninvited_users.append(user)
     return uninvited_users
 
+
 def get_invited_users(group_id: int):
     """
     Gets all user records that have been invited to a group.
@@ -85,3 +87,45 @@ def get_invited_users(group_id: int):
     for member in group.users:
         invited_users.append(member)
     return invited_users
+
+
+def get_invitations(email: str):
+    """
+    Looks at memberships matching the user with the supplied email and
+    returns all invitations that are currently active.
+
+    :param email: Email of the user being checked for invitations.
+    :returns: **List** A list of Memberships for invitations.
+    """
+    user = get_user(email)
+    invitations = []
+    for membership in user.groups:
+        if membership.invitation_accepted is False:
+            invitations.append(membership)
+    return invitations
+
+
+def get_invitation(group_id: int, email: str):
+    """
+    Looks at memberships matching the user with the supplied email and
+    returns all invitations that are currently active.
+
+    :param group_id: ID of the group to get the invitation of
+    :param email: Email of the user being checked for invitations.
+    :returns: **List** A list of Memberships for invitations.
+    """
+    invitations = get_invitations(email)
+    membership = None
+    for invitation in invitations:
+        if int(invitation.group.id) == int(group_id):
+            membership = invitation
+    return membership
+
+
+def accept_invitation(membership: Membership):
+    """
+    Accepts an invitation to a group by setting the flag for <code>invitation_accepted</code>
+     to True
+    """
+    membership.invitation_accepted = True
+    db.session.commit()
