+++ app/hindsite/group/templates/partials/accepted.html
