+++ app/hindsite/home/home_model.py
@@ -17,48 +17,6 @@ class GroupAddError(Exception):
         self.message = message
 
 
-def get_invitations(email: str):
-    """
-    Looks at memberships matching the user with the supplied email and
-    returns all invitations that are currently active.
-
-    :param email: Email of the user being checked for invitations.
-    :returns: **List** A list of Memberships for invitations.
-    """
-    user = get_user(email)
-    invitations = []
-    for membership in user.groups:
-        if membership.invitation_accepted is False:
-            invitations.append(membership)
-    return invitations
-
-
-def get_invitation(group_id: int, email: str):
-    """
-    Looks at memberships matching the user with the supplied email and
-    returns all invitations that are currently active.
-
-    :param group_id: ID of the group to get the invitation of
-    :param email: Email of the user being checked for invitations.
-    :returns: **List** A list of Memberships for invitations.
-    """
-    invitations = get_invitations(email)
-    membership = None
-    for invitation in invitations:
-        if int(invitation.group.id) == int(group_id):
-            membership = invitation
-    return membership
-
-
-def accept_invitation(membership: Membership):
-    """
-    Accepts an invitation to a group by setting the flag for <code>invitation_accepted</code>
-     to True
-    """
-    membership.invitation_accepted = True
-    db.session.commit()
-
-
 def create_group(name: str, email: str):
     """
     Creates the group and associates the group with the current user. The creating
